backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.34
mysqlclient==2.2.4
alembic==1.13.2
paho-mqtt==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.2